fix: export config at `/config` instead of `/`
Partly aligns with vike-react and vike-vue

BREAKING CHANGE: Update Vike to `0.4.160` or above.

BREAKING CHANGE: Replace `import vikeSolid from 'vike-solid'`
with `import vikeSolid from 'vike-solid/config'`. (Typically
in your `/pages/+config.h.js`.)
