+++ examples/basic/pages/+config.h.ts
@@ -1,5 +1,5 @@
 import type { Config } from "vike/types";
-import vikeSolid from "vike-solid";
+import vikeSolid from "vike-solid/config";
 
 // Default config (can be overridden by pages)
 export default {
