+++ examples/ssr-spa/pages/+config.h.ts
@@ -1,5 +1,5 @@
 import type { Config } from "vike/types";
-import vikeSolid from "vike-solid";
+import vikeSolid from "vike-solid/config";
 import Layout from "../layouts/LayoutDefault";
 import Head from "./Head";
