--- vike-solid/components/useData.tsx
@@ -1,1 +1,0 @@
-export { useData } from "../renderer/PageContextProvider";
