--- vike-solid/components/usePageContext.tsx
@@ -1,1 +1,0 @@
-export { usePageContext } from "../renderer/PageContextProvider";
