+++ vike-solid/hooks/useData.tsx
@@ -1,0 +1,14 @@
+export { useData };
+
+import { usePageContext } from "./usePageContext";
+
+/** Access `pageContext.data` from any SolidJS component
+ *
+ * See
+ * - https://vike.dev/data
+ * - https://vike.dev/pageContext-anywhere
+ */
+function useData<Data>(): Data {
+  const { data } = usePageContext() as any;
+  return data;
+}
