+++ vike-solid/main.ts
@@ -1,0 +1,5 @@
+// TODO/next-major-release: remove this file/export
+console.warn(
+  "[vike-solid][warning][deprecation] Replace `import vikeVue from 'vike-solid'` with `import vikeSolid from 'vike-solid/config'` (typically in your /pages/+config.h.js)"
+);
+export { config } from "./+config.js";
