+++ vike-solid/utils/getGlobalObject.ts
